refactor(hooks/useClipboard): change useHotkeyCopy to useClipboard for handling both hotkey and manual copy to clipboard
